Dyttet deler av koden som var i Controlleren inn i Servicen
