+++ src/main/kotlin/no/nav/klage/oppgave/api/controller/SaksbehandlerController.kt
@@ -4,12 +4,9 @@ import io.swagger.annotations.Api
 import io.swagger.annotations.ApiOperation
 import io.swagger.annotations.ApiParam
 import no.nav.klage.oppgave.api.view.Enhet
-import no.nav.klage.oppgave.api.view.Medunderskriver
 import no.nav.klage.oppgave.api.view.Medunderskrivere
 import no.nav.klage.oppgave.config.SecurityConfiguration
 import no.nav.klage.oppgave.domain.EnheterMedLovligeTemaer
-import no.nav.klage.oppgave.domain.kodeverk.Tema
-import no.nav.klage.oppgave.repositories.SaksbehandlerRepository
 import no.nav.klage.oppgave.service.SaksbehandlerService
 import no.nav.klage.oppgave.util.getLogger
 import no.nav.security.token.support.core.api.ProtectedWithClaims
@@ -20,10 +17,7 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @Api(tags = ["kabal-api"])
 @ProtectedWithClaims(issuer = SecurityConfiguration.ISSUER_AAD)
-class SaksbehandlerController(
-    private val saksbehandlerService: SaksbehandlerService,
-    private val saksbehandlerRepository: SaksbehandlerRepository
-) {
+class SaksbehandlerController(private val saksbehandlerService: SaksbehandlerService) {
 
     companion object {
         @Suppress("JAVA_CLASS_ON_COMPANION")
@@ -57,20 +51,9 @@ class SaksbehandlerController(
         @PathVariable tema: String
     ): Medunderskrivere {
         logger.debug("getMedunderskrivere is requested by $navIdent")
-        val medunderskrivere = saksbehandlerRepository.getAlleSaksbehandlerIdenter()
-            .filter { it != navIdent }
-            .filter { saksbehandlerHarTilgangTilTema(it, tema) }
-            .map { Medunderskriver(it, getNameForIdent(it)) }
-        return Medunderskrivere(tema, medunderskrivere)
+        return saksbehandlerService.getMedunderskrivere(navIdent, tema)
     }
-
-    private fun saksbehandlerHarTilgangTilTema(ident: String, tema: String) =
-        saksbehandlerRepository.getEnheterMedTemaerForSaksbehandler(ident).enheter.flatMap { it.temaer }
-            .contains(Tema.of(tema))
-
-    private fun getNameForIdent(it: String) =
-        saksbehandlerRepository.getNamesForSaksbehandlere(setOf(it)).getOrDefault(it, "Ukjent navn")
-
+    
     private fun logEnheter(enheter: List<Enhet>, navIdent: String) {
         enheter.forEach { enhet ->
             logger.debug(
